requirements.txt
fastapi
uvicorn
pymupdf
faiss-cpu
sentence-transformers
rank-bm25
ollama
langchain